feat : 일기 생성, 수정 기능

+++ src/main/java/com/todayrestarea/diary/model/DiaryRes.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 @Setter
 @AllArgsConstructor
 public class DiaryRes {
-    private int diaryIdx;
+    private Long diaryIdx;
 }
